# packages/contracts/openapi/engine.yaml
openapi: 3.1.0
info:
  title: Shakedown Engine API
  version: 0.1.5
  description: |
    > **상태: 통합 제안 (v0.1.5, 2026-10-10)**. 팀 리뷰 전이라 바뀔 수 있습니다.
    > 바꿀 때는 팀 채널에 먼저 공유하고, 버전과 `packages/contracts/README.md`의 변경 이력을 함께 고칩니다.

    엔진(apps/engine, 김도경)이 구현하고 대시보드(apps/web, 서동옥)가 호출하는 API.

    - 응답 형식은 `packages/contracts/src/index.ts`의 `Project`, `Deployment`와 같음
    - 실제 예시: `packages/contracts/fixtures/project.json`, `deployment-blocked-then-fixed.json`
    - 대시보드는 엔진 없이도 fixture로 동작함(mock 모드). 엔진이 이 명세대로 응답하면 그대로 연결됨
    - 엔진은 내부적으로 Target API(target.yaml)와 Shakedown API(shakedown.yaml)를 호출해 결과를 합침

    상태 확인 방식 (초안)
    - 상세 화면은 `events`(SSE)로 실시간 수신, 첫 화면 카드는 `GET /api/deployments/{id}`를 1초 간격으로 조회
    - 첫 화면도 나중에 SSE로 바뀔 수 있음
# 해커톤 내부 통신용이라 인증 없음. 공개 배포 시 토큰 인증 추가.
security: []
servers:
  - url: http://localhost:8700
    description: apps/engine (대시보드의 dev:web:live가 연결하는 주소)
paths:
...
  /api/deployments:
    get:
      operationId: listDeployments
      summary: 배포 목록
      parameters:
        - { name: project_id, in: query, schema: { type: string } }
      responses:
        "200":
          description: 최신순 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Deployment" }
  /api/deployments/{deployment_id}:
    get:
      operationId: getDeployment
      summary: 배포 전체 상태 (대상별 상태, 시운전 결과, 원인 보고서, AI 비용)
      parameters:
        - { name: deployment_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 배포
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
...
components:
  schemas:
...
    Deployment:
      type: object
      description: "전체 필드는 packages/contracts/src/index.ts의 Deployment"
      required: [id, project_id, status, targets, attempts]
      properties:
        id: { type: string }
        project_id: { type: string }
        status:
          type: string
          enum: [queued, building, deploying, shakedown, analyzing, fixing, deployed, promoted, warned, blocked, failed]
          description: "deployed·promoted·warned·blocked·failed에서 엔진은 스스로 더 진행하지 않음. blocked는 POST /fix로 한 번 fixing으로 재개될 수 있음"
        mode: { type: string, enum: [comparison] }
        shakedown_id: { type: string }
        release_gate: { type: string, enum: [passed, review, blocked] }
        traffic_blocked: { type: boolean, description: "관리 AWS DELETE 차단 확인 시 true. 기존 외부 URL 트래픽은 차단하지 않음." }
        targets:
          type: object
          description: "대상 이름 → 상태. 기존 환경은 external(관리/배포 완료 아님), 정리 완료는 stopped, 관리 대상은 TargetDeployment 필드 사용. 관리 대상의 deployment_id는 Target API에 실제로 쓴 ID(수정 재배포 뒤에는 새 ID), request는 받아들여진 Target API 본문(deployment_id 제외)"
          additionalProperties: true
        attempts:
          type: array
          description: "시운전 회차. 각 회차의 steps/verdict/report는 shakedown.yaml의 Shakedown과 같은 형식. applied_fix는 그 회차가 차단된 뒤 적용한 수정안(POST /fix)"
          items: { type: object, additionalProperties: true }
        timings: { type: object, additionalProperties: { type: number } }
        ai_cost: { type: object, additionalProperties: true }
...